# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cms/categories:
    get:
      operationId: listCmsCategories
      tags:
        - CMS Categories
      responses:
        '200':
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      operationId: createCmsCategory
      tags:
        - CMS Categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertCategoryBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /cms/categories/{id}:
    patch:
      operationId: updateCmsCategory
      tags:
        - CMS Categories
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertCategoryBody'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteCmsCategory
      tags:
        - CMS Categories
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /cms/tags:
    get:
      operationId: listCmsTags
      tags:
        - CMS Tags
      responses:
        '200':
          description: Tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
    post:
      operationId: createCmsTag
      tags:
        - CMS Tags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertTagBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
  /cms/tags/{id}:
    patch:
      operationId: updateCmsTag
      tags:
        - CMS Tags
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertTagBody'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
    delete:
      operationId: deleteCmsTag
      tags:
        - CMS Tags
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
...
  /cms/users/{id}/roles:
    put:
      operationId: setCmsUserRoles
      tags:
        - CMS Users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roles
              properties:
                roles:
                  type: array
                  items:
                    $ref: '#/components/schemas/RoleName'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CmsUser'
  /insights:
    get:
      operationId: listInsights
      tags:
        - Insights
      parameters:
        - in: query
          name: categorySlug
          schema:
            type: string
        - in: query
          name: tagSlug
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 12
      responses:
        '200':
          description: Public posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicPostListResponse'
  /insights/{slug}:
    get:
      operationId: getInsight
      tags:
        - Insights
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Public post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicPost'
        '404':
          $ref: '#/components/responses/NotFound'
  /insights/{slug}/comments:
    get:
      operationId: listInsightComments
      tags:
        - Insights
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Approved comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicComment'
    post:
      operationId: submitInsightComment
      tags:
        - Insights
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitCommentBody'
      responses:
        '202':
          description: Accepted (pending moderation)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitCommentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        details: {}
    RoleName:
      type: string
      enum:
        - admin
        - editor
        - author
        - contributor
...
    CommentStatus:
      type: string
      enum:
        - pending
        - approved
        - spam
        - deleted
...
    CmsUser:
      type: object
      required:
        - id
        - clerkUserId
        - roles
      properties:
        id:
          type: string
          format: uuid
        clerkUserId:
          type: string
        email:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleName'
        createdAt:
          type: string
          format: date-time
    Category:
      type: object
      required:
        - id
        - slug
        - name
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
    UpsertCategoryBody:
      type: object
      required:
        - slug
        - name
      properties:
        slug:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
    Tag:
      type: object
      required:
        - id
        - slug
        - name
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
    UpsertTagBody:
      type: object
      required:
        - slug
        - name
      properties:
        slug:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
...
    AuthorSummary:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
...
    PublicPost:
      type: object
      required:
        - id
        - slug
        - title
        - publishedAt
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        excerpt:
          type: string
          nullable: true
        bodyHtml:
          type: string
          nullable: true
        heroImageUrl:
          type: string
          nullable: true
        ogImageUrl:
          type: string
          nullable: true
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
        seoCanonicalUrl:
          type: string
          nullable: true
        readingTimeMin:
          type: integer
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        author:
          $ref: '#/components/schemas/AuthorSummary'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
    PublicPostListResponse:
      type: object
      required:
        - items
        - page
        - pageSize
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PublicPost'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    PublicComment:
      type: object
      required:
        - id
        - authorName
        - bodyText
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        parentCommentId:
          type: string
          format: uuid
          nullable: true
        authorName:
          type: string
        bodyText:
          type: string
        createdAt:
          type: string
          format: date-time
    SubmitCommentBody:
      type: object
      required:
        - authorName
        - authorEmail
        - bodyText
      properties:
        authorName:
          type: string
          minLength: 1
          maxLength: 120
        authorEmail:
          type: string
          format: email
        bodyText:
          type: string
          minLength: 1
          maxLength: 5000
        parentCommentId:
          type: string
          format: uuid
          nullable: true
    SubmitCommentResponse:
      type: object
      required:
        - id
        - status
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/CommentStatus'
...