# lib/api-zod/src/generated/types/submitCommentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitCommentBody {
  /**
   * @minLength 1
   * @maxLength 120
   */
  authorName: string;
  authorEmail: string;
  /**
   * @minLength 1
   * @maxLength 5000
   */
  bodyText: string;
  parentCommentId?: string | null;
}